# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zanichelli API IdentityProvider",
        "description": "REST APIs to use SSO (single-sign-on) for applications ",
        "contact": {
            "name": "Zanichelli DEV team",
            "email": "gsiciliano@zanichelli.it"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/users/{id}/user-roles": {
            "get": {
                "tags": [
                    "User-Role"
                ],
                "summary": "Retrieve role on department of user",
                "description": "Retrieve role on department of user",
                "operationId": "UserRole.retrieve-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User-Role"
                ],
                "summary": "Assigne role on department to user",
                "description": "__*Security:*__ __*can be used only by clients with 'manager' role*__",
                "operationId": "UserRole.create",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "department_id": {
                                            "description": "Id of department",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "role_id": {
                                            "description": "Id of role",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "204": {
                        "description": "Operation successfull, user already has role passed",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Invalid scope or client role, Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/user-role/{id}": {
            "delete": {
                "tags": [
                    "User-Role"
                ],
                "summary": "Remove role on department to user by user-role-id",
                "description": "__*Security:*__ __*can be used only by clients with 'manager' role*__",
                "operationId": "UserRole.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User-role id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Error on saving",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Invalid scope or client role, Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "http://localhost:8083/oauth/token",
                        "scopes": []
                    }
                },
                "scheme": [
                    "http",
                    "https"
                ]
            }
        }
    },
...
}